azure-function/requirements.txt
azure-functions
azure-storage-blob
pandas
numpy
